ToDo app with JS
Empty characters cannot be added to the list. Additionally, the same error occurs when nothing is written.
The JavaScript code you will write should include a function for adding elements, a function for deleting elements, and a function for marking tasks as done.
When an element is added or an error occurs, you may have noticed a notification at the top right. This is provided by a Bootstrap Toast notification