Asian Kitchen Menu Design with JS
The site is designed as a Far East restaurant menu and we can navigate between pages thanks to JavaScript.

The content on the site will be defined as objects in a menu array in JavaScript.
There will be id, title, category, price, img and desc sections.
You can change the content as you wish, but extra attention was paid to the category names while changing.
The buttons also came from the js file.