added Mutations
Create, update, delete and delete all Mutations related to all types were prepared.

A Mutation that will add a new User was written.
A Mutation that will update a User was written.
A Mutation that will delete a User was written.
A Mutation that will delete all Users was written.
The above items were also applied to Event, Location and Participant.